Update report to use order repository instead of openning connection by itself
