+++ backend/online-shopping-webstore/src/main/java/edu/mum/onlineshopping/util/EmailUtil.java
@@ -1,0 +1,76 @@
+package edu.mum.onlineshopping.util;
+
+import java.io.File;
+import java.io.IOException;
+
+import java.util.Date;
+import java.util.Properties;
+import javax.mail.*;
+import javax.mail.internet.AddressException;
+import javax.mail.internet.InternetAddress;
+import javax.mail.internet.MimeBodyPart;
+import javax.mail.internet.MimeMessage;
+import javax.mail.internet.MimeMultipart;
+
+public class EmailUtil {
+    public static void sendEmailWithAttachments(String host, String port,
+                                                final String userName, final String password, String toAddress,
+                                                String subject, String message, File out)
+            throws AddressException, MessagingException {
+        // sets SMTP server properties
+        Properties properties = new Properties();
+        properties.put("mail.smtp.host", host);
+        properties.put("mail.smtp.port", port);
+        properties.put("mail.smtp.auth", "true");
+        properties.put("mail.smtp.starttls.enable", "true");
+        properties.put("mail.user", userName);
+        properties.put("mail.password", password);
+
+        // creates a new session with an authenticator
+        Authenticator auth = new Authenticator() {
+            public PasswordAuthentication getPasswordAuthentication() {
+                return new PasswordAuthentication(userName, password);
+            }
+        };
+        Session session = Session.getInstance(properties, auth);
+
+        // creates a new e-mail message
+        Message msg = new MimeMessage(session);
+
+        msg.setFrom(new InternetAddress(userName));
+        InternetAddress[] toAddresses = { new InternetAddress(toAddress) };
+        msg.setRecipients(Message.RecipientType.TO, toAddresses);
+        msg.setSubject(subject);
+        msg.setSentDate(new Date());
+
+        // creates message part
+        MimeBodyPart messageBodyPart = new MimeBodyPart();
+        messageBodyPart.setContent(message, "text/html");
+
+        // creates multi-part
+        Multipart multipart = new MimeMultipart();
+        multipart.addBodyPart(messageBodyPart);
+
+        // adds attachments
+        if (out != null ) {
+
+            MimeBodyPart attachPart = new MimeBodyPart();
+
+            try {
+                attachPart.attachFile(out);
+            } catch (IOException ex) {
+                ex.printStackTrace();
+            }
+
+            multipart.addBodyPart(attachPart);
+
+        }
+
+        // sets the multi-part as e-mail's content
+        msg.setContent(multipart);
+
+        // sends the e-mail
+        Transport.send(msg);
+
+    }
+}
